Implements in-game inventory and menu system
Introduces a new in-game inventory system accessible via a menu.

This commit adds prefabs for the inventory and main menu, and implements scene management for dynamic UI loading. It also refactors menu handling and introduces a scene manager to support modal scene loading, enhancing UI modularity and in-game usability.

Adds multi-player menu.
Updates single-player menu item UI

+++ src/ModLoader/handler/IMenuManager.cs
@@ -3,6 +3,6 @@
 namespace ModLoader.handler;
 
 public interface IMenuManager {
-    public void AddMenuGroup(string id, short order = -1);
-    public void AddMenuItem(string groupId, string itemId, GetString itemName, short order, GetString description, Action action);
+    public void AddMenuGroup(string id, int order = -1);
+    public void AddMenuItem(string groupId, string itemId, GetString itemName, int order, GetString description, Action action);
 }
